# Execute multi-row INSERTs sequentially