This PR changes multi-row INSERTs to use sequential execution, over a single connection per worker. The current approach uses the multi-shard execution infrastructure for parallel DDL, but this causes poor performance (see #1604) due to the high number of connections.

Latency-wise, sequential execution is not the optimal way to execute a multi-row INSERT, but it gives substantially lower latency and higher throughput than the current approach.

Before:

1 shard: 1900 tps = 190,000 rows/sec (17ms avg)
2 shards: 246 tps = 24,600 rows/sec (131ms avg)
4 shards: 175 tps = 17,500 rows/sec (182ms avg)
8 shards: 132 tps = 13,200 rows/sec (240ms avg)
16 shards: 92 tps = 9,200 rows/sec (351ms avg)
32 shards: 58 tps = 5,800 rows/sec (555ms avg)

After:

1 shard: 2081 tps = 208,100 rows/sec (15ms avg)
2 shards: 1484 tps = 148,400 rows/sec (22ms avg)
4 shards: 1221 tps = 122,100 rows/sec (26ms avg)
8 shards: 893 tps = 89,300 rows/sec (36ms avg)
16 shards: 622 tps = 62,200 rows/sec (51ms avg)
32 shards: 385 tps = 38,500 rows/sec (83ms avg)