@sumedhpathak 2017-08-22
After chatting with @ozgune we decided to include this in 7.0 as well, given the performance benefits. @jasonmp85 Could you review the change?